pyproject.toml
[project]
name = "fastapi-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "celery[redis]>=5.6.0",
    "fastapi[standard]>=0.124.0",
    "pillow>=12.0.0",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
    "gliner>=0.2.24",
]
